LayoutTests/fast/mediastream/MediaStreamTrack-getSettings-expected.txt

Tests MediaStreamTrack.getSettings.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


video track settings:
  settings.width = 640
  settings.height = 480
  settings.frameRate = 30
  settings.facingMode = user
  settings.deviceId = <UUID>

audio track settings:
  settings.volume = 1
  settings.echoCancellation = false
  settings.deviceId = <UUID>

According to the spec: "[every setting] MUST be a member of the set defined for that property by getCapabilities()"

PASS "width" in track.getCapabilities() is true
PASS "height" in track.getCapabilities() is true
PASS "frameRate" in track.getCapabilities() is true
PASS "facingMode" in track.getCapabilities() is true
PASS "deviceId" in track.getCapabilities() is true
PASS "volume" in track.getCapabilities() is true
PASS "echoCancellation" in track.getCapabilities() is true
PASS "deviceId" in track.getCapabilities() is true
PASS successfullyParsed is true

TEST COMPLETE
